refactor(auth): migrate useAuthSettings to queryOptions per-domain pattern
Creates src/queries/auth.queries.ts with queryOptions (TanStack Query v5
recommended pattern). Query key and queryFn are co-located, eliminating
the hardcoded ['auth', 'settings'] antipattern.

useAuthSettings now consumes authQueries.settings() — enabled flag
preserved, behavior unchanged.
